+++ web/themes/quartiers_solidaires/assets/components/molecules/form-control/form-ajax.js
@@ -1,10 +1,48 @@
 const formAjax = () => {
-  (function ($) {
+  (function ($, Drupal) {
+
+    // Handle buttons with confirmation step
+    let timer = null;
+    const delay = 2500;
+
+    $('form[data-ajax="true"]').on('click', '[type="submit"][data-confirm]', function(e) {
+      const $this = $(this);
+      const text = $this.text();
+      const $icon = $this.find('.icon').clone();
+      const confirmText = $this.data('confirm');
+console.log($icon);
+      // Send request if we are in pending state
+      if ($this.data('pending')) {
+        window.clearTimeout(timer);
+        $this
+          .text(Drupal.t('qs.form_confirmed_feedback'))
+          .removeData('pending');
+        return true;
+      }
+
+      $this
+        .text(confirmText)
+        .prepend($icon)
+        .data('confirm', text)
+        .data('pending', 'true')
+        .addClass('btn-confirm');
+
+      timer = window.setTimeout(function() {
+        $this
+          .text(text)
+          .prepend($icon)
+          .data('confirm', confirmText)
+          .removeData('pending')
+          .removeClass('btn-confirm');
+      }, delay);
+
+      return false;
+    });
 
     $(document).on('submit', 'form[data-ajax="true"]', function(e) {
       e.preventDefault();
       const $this = $(this);
-
+      
       $.ajax({
           type: $this.attr('method'),
           url:  $this.attr('action'),
@@ -30,7 +68,7 @@ const formAjax = () => {
           },
       });
     });
-  })(jQuery);
+  })(jQuery, Drupal);
 };
 
 export default formAjax;